+++ CHANGELOG.md
@@ -6,12 +6,14 @@
 
 - Option: `Include Below Average Equipment`
 - Option: `Include Unsuitable Equipment`
+- Command: `/dejunk keybinds`
 
 ### Changed
 
 - SavedVariables now populate/depopulate default values on login/logout
 - Made some minor UI modifications
 - Junk frame now displays individual item stack prices
+- Updated the options frame to have "Keybinds" button
 
 ### Fixed